cli/pyproject.toml
[project]
name = "harness"
version = "0.1.0"
description = "Workspace Harness CLI — bootstrap, context, policy, exec, report."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
dependencies = [
    "click>=8.1",
    "jinja2>=3.1",
    "pyyaml>=6.0",
    "rich>=13.7",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0",
]

[project.scripts]
harness = "harness.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/harness"]

[tool.hatch.build.targets.wheel.force-include]
"src/harness/templates" = "harness/templates"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
